# docsclear/swagger.yaml
basePath: /
definitions:
...
  entity.ProductTontineSettings:
    properties:
      allow_commercial_circle:
        type: boolean
      allow_corporate_circle:
        type: boolean
      allow_family_circle:
        type: boolean
      created_at:
        type: string
      is_tontine_enabled:
        type: boolean
      max_participants:
        type: integer
      min_participants:
        type: integer
      product_id:
        type: string
      shop_id:
        type: string
      updated_at:
        type: string
    type: object
...
  shophandler.TontineSettingsRequest:
    properties:
      allow_commercial_circle:
        example: true
        type: boolean
      allow_corporate_circle:
        example: true
        type: boolean
      allow_family_circle:
        example: true
        type: boolean
      is_tontine_enabled:
        example: true
        type: boolean
      max_participants:
        example: 12
        type: integer
      min_participants:
        example: 4
        type: integer
      product_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  tontineusecase.CreateGroupRequest:
    properties:
      circle_type:
        type: string
      creator_customer_id:
        description: Optionnel si créé par marchand
        type: string
      product_id:
        type: string
      total_cycles:
        type: integer
    type: object
  tontineusecase.JoinGroupRequest:
    properties:
      customer_id:
        type: string
      invite_code:
        type: string
    type: object
  tontineusecase.PayCycleRequest:
    properties:
      customer_id:
        type: string
      flow:
        description: indirect (défaut) ou direct
        type: string
      group_id:
        type: string
      operator:
        description: orange_money, moov_money, telecel, etc.
        type: string
      phone_number:
        description: +226...
        type: string
    type: object
...
  utils.AppError:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@goshop.dev
    name: API Support
  description: GoShop - E-commerce API with observability and metrics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GoShop API
  version: 1.0.0
paths:
...
  /api/shops/{id}/tontine-settings:
    get:
      consumes:
      - application/json
      description: Retourne la configuration de tontine actuelle pour un produit spécifique
        dans une boutique donnée.
      parameters:
      - description: ID de la boutique (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: ID du produit (UUID)
        in: query
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ProductTontineSettings'
        "400":
          description: ID de boutique invalide ou product_id manquant
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "403":
          description: Interdit (vous n'êtes pas le propriétaire)
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Boutique ou paramètres tontine introuvables
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Récupérer les paramètres de tontine d'un produit
      tags:
      - Tontine Settings
    put:
      consumes:
      - application/json
      description: Active ou désactive la tontine pour un produit et configure les
        paramètres (cercles autorisés, nombre de participants, etc.).
      parameters:
      - description: ID de la boutique (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Nouveaux paramètres de tontine
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shophandler.TontineSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ProductTontineSettings'
        "400":
          description: Payload invalide, ID invalide ou validation échouée
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "403":
          description: Interdit (vous n'êtes pas le propriétaire)
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Boutique introuvable
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Mettre à jour les paramètres de tontine d'un produit
      tags:
      - Tontine Settings
...
  /api/tontine/groups:
    post:
      consumes:
      - application/json
      description: Initialise un nouveau groupe de tontine pour un produit spécifique.
      parameters:
      - description: Détails du groupe à créer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tontineusecase.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Payload invalide ou règles de tontine non respectées
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "403":
          description: Interdit (KYC non validé ou droits insuffisants)
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Créer un groupe de tontine
      tags:
      - Tontine
  /api/tontine/groups/{group_id}/pay:
    post:
      consumes:
      - application/json
      description: Lance le processus de paiement pour la cotisation d'un cycle spécifique.
      parameters:
      - description: ID du groupe de tontine (UUID)
        in: path
        name: group_id
        required: true
        type: string
      - description: Détails du paiement (provider, phone_number, etc.)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tontineusecase.PayCycleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Payload invalide ou cycle non éligible
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Groupe ou cycle introuvable
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Payer une cotisation de tontine
      tags:
      - Tontine
  /api/tontine/groups/{group_id}/payments:
    get:
      consumes:
      - application/json
      description: Retourne l'historique des cotisations payées par un client spécifique
        dans un groupe de tontine.
      parameters:
      - description: ID du groupe de tontine (UUID)
        in: path
        name: group_id
        required: true
        type: string
      - description: ID du client (UUID)
        in: query
        name: customer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Paramètres manquants
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "403":
          description: Interdit (accès aux données d'un autre client)
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Groupe introuvable
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Lister les paiements d'un client dans un groupe
      tags:
      - Tontine
  /api/tontine/groups/join:
    post:
      consumes:
      - application/json
      description: Permet à un client de rejoindre un groupe de tontine existant via
        un code d'invitation.
      parameters:
      - description: Code d'invitation et ID client
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tontineusecase.JoinGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Payload invalide ou code incorrect
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "403":
          description: Interdit (KYC non validé)
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Groupe introuvable
          schema:
            $ref: '#/definitions/utils.AppError'
        "409":
          description: Groupe déjà complet ou client déjà membre
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Rejoindre un groupe de tontine
      tags:
      - Tontine
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"